requirements.txt
clear==1.0.1
Django==2.2.1
django-multiselectfield==0.1.11
goodlogging==1.0.1
psycopg2-binary==2.8.3
pytz==2019.3
rarfile==2.7
requests==2.6.0
sqlparse==0.3.0
